+++ lib/ipdb_decoder/tire.ex
@@ -1,17 +1,18 @@
 defmodule IPDBDecoder.Tire do
   @moduledoc false
 
+  alias IPDBDecoder.Database
   use Bitwise
 
-  def find_v4node(database) do
+  def find_v4node(database, options \\ []) do
     bits = List.duplicate(0, 80) ++ List.duplicate(1, 16)
 
-    case find(database, bits) do
+    case find(database, bits, options) do
       {:stop, current} -> current
     end
   end
 
-  def find(database, ip) when is_tuple(ip) do
+  def find(database, ip, options) when is_tuple(ip) do
     offset = 1 <<< (tuple_size(ip) * 2)
 
     bits =
@@ -20,45 +21,58 @@ defmodule IPDBDecoder.Tire do
       |> Enum.map(fn v -> tl(Integer.digits(v + offset, 2)) end)
       |> List.flatten()
 
-    case find(database, bits) do
+    case find(database, bits, options) do
       {:stop, _} -> {:error, :not_found_ip}
       {:ok, _} = ok -> ok
       {:error, _} = error -> error
     end
   end
 
-  def find(database, bits) when is_list(bits) do
+  def find(database, bits, options) when is_list(bits) do
     case Enum.count(bits) do
       32 ->
-        find_next(database, bits, database.v4node)
+        find_next(database, bits, database.v4node, options)
 
       _ ->
-        find_next(database, bits, 0)
+        find_next(database, bits, 0, options)
     end
   end
 
-  def find_next(database, [head | tail], current) do
-    %{meta: %{"node_count" => total}} = database
+  def find_next(database, [head | tail], current, options) do
+    %{meta: %{:node_count => total}} = database
 
     case binary_part(database.chunks, current <<< 3 ||| head <<< 2, 4) do
       <<next::unsigned-32>> when next == total ->
         {:error, :not_found_ip}
 
       <<next::unsigned-32>> when next > total ->
-        find_data(database.chunks, next - total + total * 8)
+        find_data(database, next - total + total * 8, options)
 
       <<next::unsigned-integer-32>> ->
-        find_next(database, tail, next)
+        find_next(database, tail, next, options)
     end
   end
 
-  def find_next(_, [], current) do
+  def find_next(_, [], current, _) do
     {:stop, current}
   end
 
-  def find_data(chunks, postion) do
-    <<offset::integer-16>> = binary_part(chunks, postion, 2)
-    info = binary_part(chunks, postion + 2, offset)
-    {:ok, String.split(info, "\t")}
+  def find_data(database, postion, options) do
+    <<offset::integer-16>> = binary_part(database.chunks, postion, 2)
+    info = binary_part(database.chunks, postion + 2, offset)
+    range = Database.slice_range(database, Keyword.get(options, :locale))
+
+    value =
+      info
+      |> String.split("\t")
+      |> Enum.slice(range)
+
+    value =
+      [database.meta[:fields], value]
+      |> List.zip()
+      |> Enum.map(fn {key, value} -> {String.to_atom(key), value} end)
+      |> Map.new()
+
+    {:ok, value}
   end
 end