+++ test/ipdb_decoder_test.exs
@@ -3,17 +3,23 @@ defmodule IPDBDecoderTest do
   use ExUnit.Case
 
   @path Path.join([__DIR__, "../priv/ipipfree.ipdb"])
+  @test_ip "202.112.84.101"
+  @result %{
+    :city_name => "北京",
+    :country_name => "中国",
+    :region_name => "北京"
+  }
 
   test "parse_database_file" do
     database = @path |> IPDBDecoder.parse_database_file()
-    {:ok, result} = IPDBDecoder.pipe_lookup(database, "127.0.0.1")
-    assert result == ["本机地址", "本机地址", ""]
+    {:ok, result} = IPDBDecoder.pipe_lookup(database, @test_ip)
+    assert result == @result
   end
 
   test "parse_database" do
     database = @path |> File.read!() |> IPDBDecoder.parse_database()
-    {:ok, result} = IPDBDecoder.pipe_lookup(database, "127.0.0.1")
-    assert result == ["本机地址", "本机地址", ""]
+    {:ok, result} = IPDBDecoder.pipe_lookup(database, @test_ip)
+    assert result == @result
   end
 
   test "parse_database_nofile" do
@@ -21,22 +27,22 @@ defmodule IPDBDecoderTest do
 
     assert {:error, :open_file_failed} ==
              IPDBDecoder.parse_database_file("wrongfile.ipdb")
-             |> IPDBDecoder.pipe_lookup("127.0.0.1")
+             |> IPDBDecoder.pipe_lookup(@test_ip)
   end
 
   test "lookup" do
     {:ok, database} = @path |> IPDBDecoder.parse_database_file()
-    {:ok, result} = IPDBDecoder.lookup(database, "127.0.0.1")
-    assert result == ["本机地址", "本机地址", ""]
+    {:ok, result} = IPDBDecoder.lookup(database, @test_ip)
+    assert result == @result
   end
 
   test "lookup_pipe" do
     {:ok, result} =
       @path
       |> IPDBDecoder.parse_database_file()
-      |> IPDBDecoder.pipe_lookup("127.0.0.1")
+      |> IPDBDecoder.pipe_lookup(@test_ip)
 
-    assert result == ["本机地址", "本机地址", ""]
+    assert result == @result
   end
 
   test "lookup wrong ip" do