+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## 0.0.2 (June 25, 2019)
+  - parse result to map and support locale if database provided.
+
 ## 0.0.1 (June 25, 2019)
 
-- init version of IPIP.net geoip database decoder and publish it to Hex with ❤️
+  - init version of IPIP.net geoip database decoder and publish it to Hex with ❤️